memory-agent/requirements.txt
pydantic>=2.0
langchain
langgraph
langchain-groq
chromadb
sentence-transformers
python-dotenv
